requirements.txt
tensorflow-cpu==2.6
Pillow
google.cloud